[LayoutNG] Add PropagateChildBreakValues flag to the ConstraintSpace
This CL adds a new flag to the ConstraintSpace to propagate child break
values (initial-break-before and previous-break-after).

The flag is currently set when setting the fragmentation type. However,
it will also be used in future changes outside of the normal
fragmentation pass to determine the initial-break-before and
final-break-after for rows.

The code that performs child break value propagation in
MovePastBreakpoint() was moved to the builder, and it is run
whenever the new flag is set to true.

This fixed a bug where we failed to propagate up the break values of
a child if there was no BFC block-offset. The reason was that we
never called BreakBeforeChildIfNeeded() in that case. By moving the
propagation to the builder, instead, we can ensure the break values
are correctly propagated in this case, as well.

Bug: 829028,1293291
Change-Id: I2afd288cad99d11a1c414aca5844d7a77afef46f