bluetooth: Fix crash in invoking some APIs when the context is gone
Fix the failure/crash when invoking APIs exposed by the bluetooth module
while the context is gone. As mojo service has the mechanism to detect
context existence and set service bounded state accordingly, relying on
checking at service bounded state and throwing an exception when the
remote service is not bounded.

Bug: 1282963

Change-Id: I6b6b2ee2724c7eb73bcd29b5d5c2fbdbaf325683